ai-service/requirements.txt
# LiveKit - Real-time room participant
livekit>=0.13.0
livekit-api>=0.6.0

# Transcription - Streaming API
deepgram-sdk>=3.7.0,<4.0.0

# AI Notes Generation
anthropic>=0.39.0

# Utilities
python-dotenv>=1.0.0
websockets>=12.0
aiohttp>=3.9.0
numpy>=1.26.0
